refactor(jazz-tools): Phase 7 — extract graph tests, set up graph/ submodule layout
graph.rs was 4.6K LOC, ~1.5K of which were tests at the bottom of the file.
This PR:

  - moves graph.rs → graph/mod.rs, creating the directory the spec
    targets for the eventual compile/execute split
  - extracts the trailing #[cfg(test)] mod tests block (~1500 LOC) into
    graph/tests.rs

Result: graph/mod.rs is ~3.1K LOC of source (down from 4.6K), with the
test scaffolding split off. The compile/execute split called for in the
spec is deferred to a follow-up — the impl QueryGraph block (~2.4K of
the remaining mod.rs) is one tightly coupled unit, and splitting it
across submodules requires bumping field/method visibility throughout
to pub(super). That's a larger, riskier surgery and worth its own PR.

Spec: specs/todo/c_later/crate_structure_cleanup.md (Phase 7 — partial;
graph/mod.rs and tests.rs landed, graph_compile.rs / graph_execute.rs
follow-up pending).
